Stop creating linter report files in CI
Currently, when one of the linters reports a problem, if you want to see
what that problem is, you have to do the following:

1. Click on the "Details" link in the PR.
2. Click "Summary".
3. Click the artifact (also, depending on the browser, choose where to save it).
4. Open the downloaded file (which is an archive).
5. Open the report file from the archive.
6. Delete the archive after you're done.

That is far too much hassle, and all it gives you is a slightly fancier-
looking report (and sometimes not even that, like in the case of black).

So just remove all the report building and uploading. Now the above
procedure is reduced to:

1. Click "Details".
2. Expand the appropriate build step.