Remove --no-binary=av from requirements files
Since c56f2e9a, we use this flag in the Dockerfile, so it's not necessary to
duplicate it in the requirements files. Omitting it makes it easier to set
up development environments, since developers don't need to install FFmpeg.

This also fixes pip-compile runs, which were broken, because PyAV is
incompatible with Cython 3. Now pip-compile can use the precompiled version.