Fix accidental prebuilt PyAV usage in the Dockerfile (#6573)
When I integrated pip-compile, I neglected to notice that in the
generated `*.txt` files the --no-binary=av option is on its own line,
and therefore is stripped away by the sed command.

Hardcode this option in the Dockerfile to make sure we build PyAV from
source. Also, add a workaround for the fact that PyAV is incompatible
with the recently-released Cython 3.