YOLOv7: use the correct version of ONNX Runtime for GPU
Currently, we install plain `onnxruntime`, which means that when the function
is deployed with `deploy_gpu.sh`, it ends up using the CPU anyway.

`onnxruntime-gpu` depends on a specific version of CUDA, so update the base
image and bound the version of the package to avoid mismatches in the
future.