Remove unnecessary COUNT for labels (#6918)
Optimized /api/projects, /api/tasks, and /api/jobs requests. Previously
for unification, we decided to add a count of labels into every response
for each project, task, and job. It leads to performance issues. UI and
probably other clients don't use the information because it is useless.
Removing it, helps to optimize the SQL request several times.

---------

Co-authored-by: Boris Sekachev <boris@cvat.ai>