Add Regal for linting Rego
Hello CVAT friends! 👋😃

This PR introduces [Regal](https://github.com/styrainc/regal)) for linting
the Rego included in this project. The policy code here is generally in a
really good shape, so the linter did not have too much to do. Good work on that!

Violations of a few rules were fixed as part of this PR, as they do not change
the semantics of evaluation:

* [line-length](https://docs.styra.com/regal/rules/style/line-length)
* [unconditional-assignment](https://docs.styra.com/regal/rules/style/unconditional-assignment)
* [use-assignment-operator](https://docs.styra.com/regal/rules/style/use-assignment-operator)

A few other ones were ignored using Regal's configuration file. Whether
some of them should be fixed in the future I'll leave up to the project's
maintainers.

I've also added Regal as a build workflow, along with some documentation.
Let me know if I missed anything!

Signed-off-by: Anders Eknert <anders@styra.com>