Support running CVAT with an external database via Docker Compose (#7055)
This would help users who don't want or need the complexity of
Kubernetes, but would still like to use an external database.

With Docker Compose, you can initialize a secret from an environment
variable, but you can't load a secret _into_ an environment variable. So
in order to be able to read the DB password from a secret, I had to
introduce a new `CVAT_POSTGRES_PASSWORD_FILE` variable.