Turn on Traefik access logs
This will be useful for debugging issues and gathering statistics.

Use the JSON format, since that allows us to include more information than
the traditional HTTP access logs. Switch the other logs to JSON as well, for
ease of parsing.

Add a cap on the size of the log in Docker compose, since the JSON logs are
pretty bulky, and we don't want users to inadvertently run out of space.