Helm: keep the data PersistentVolumeClaim after uninstall (#7123)
Currently, Helm will delete the PVC when you uninstall your CVAT
release, which makes it easy to unintentionally lose data. Add a
`helm.sh/resource-policy` annotation to prevent that from happening.

This makes the data PVC consistent with the other PVCs created as part
of the Helm chart (ones belonging to PostgreSQL, KeyDB and Clickhouse),
which are also not deleted upon uninstall, albeit for a different
reason: they are created by the corresponding StatefulSets, and
StatefulSets retain their volumes when they are deleted.