Fix compatibility with older versions of Docker Compose
In 6ae1cffd, I accidentally left behind the empty `command:` key. This
works, but only with Docker Compose v2.17.1+. Since we don't actually need
this key, we might as well remove it and restore compatibility with older
versions.

Also, rearrange the commented out code so that you only need to uncomment
one block to enable the Traefik dashboard.