Use separate services for storing job queues and cache
These types of data have different characteristics and we have different
expectations on them:

* job queues are small and we'd rather not lose them (although losing
  them is not fatal);

* cached chunks are large and we don't care if we lose them.

We currently store both in KeyDB, which has shown itself to not be
especially reliable. A few times we've had to clear the KeyDB store due to
data corruption, which destroyed the queues as well. While we'll probably
end up replacing KeyDB with something else, it would still be useful to have
the ability to just clear the cache volume without taking out the job queues
in the process.

As a solution to this, add a Redis service to be used only for the queues
(and potentially for other small data items). Using the original Redis
instead of KeyDB should also help with reliability (at least as far as the
job queues are concerned).