Update Clickhouse to the latest version
The main reason for this is to update the Helm chart, because the current
one conflicts with the Redis chart I want to add.

But also the version we have is quite old and out of support, so it's due
for an update anyway.

The change in `objects.py` is due to the fact that with the new version of
Clickhouse, the value of `date` somehow became a timezone-aware datetime
object, so the output of `isoparse` now includes a UTC offset, and appending
a "Z" now creates a malformed date. I chose to replace this with a custom
format string, which is consistent with other date formatting code in the
`analytics_report` app.