Removed `PermissionDenied` error thrown before OPA call (#7259)
At first, this part of the code was an optimization. If we know that a user is
not a member but tries to access org resources we throw an error. But
during the development of other features(like invitations to organization)
we need to give user an access to such resources. That leads to
rewriting this part of the code with lots of `if` statements.

Generally, all permission decisions should be made by OPA.