Use OPA from the Docker image for testing
Currently, the testing pipeline downloads a copy of OPA from the official
site. But we already have a perfectly usable copy in the Docker image that
we download for the Django unit tests. So just use that copy instead.

This reduces the amount of downloading the pipeline has to do, and eliminates
a hardcoded version number.