Fix unhandled Redis exception in the /api/server/health/ endpoint
`CustomCacheBackend` does not handle `RedisError`, so if Redis is not
responding, the backend will propagate the exception, and the endpoint will
crash with a 500 response code.

This problem has been fixed upstream in django-health-check 3.18, so just
update to the latest version and remove our custom cache backend.

The upstream version, unlike ours, doesn't handle `sqlite3.DatabaseError`,
but we're not using SQLite for caching anymore, so I don't think it matters.