Strip useless linebreak in the payload of send:exception events (#7407)
`traceback.format_exception` adds a `\n` to each element of the
resulting list, so the message will always have a pointless `\n` at the
end. Removing it makes CSV reports that include the message easier to
parse visually, because there's no longer a linebreak in the middle of
each line.