Defer most signal handling code until after the end of the current transaction
Non-database actions like deleting directories, sending webhooks, scheduling
reports should only be done after the current transaction is committed. If
we do it immediately, and the transaction is later aborted, then we will
have (for example) sent a webhook about an event that didn't actually
happen.

There's also a secondary benefit to moving this action outside of the
transaction; the less time we spend inside a transaction, the better, since
a transaction may lock out other clients from working on the affected DB
rows.

In addition, prevent the affected actions from crashing the view handler
with an exception (using the `robust=True` option). I don't think it's
reasonable to (for example) return a 500 response to a `PATCH` request just
because we failed to send the corresponding webhook.

There is one more type of action that should be modified in this way
(sending events), but it would be easier to do that after a refactoring
that I did in another patch, so I'll do it later.