Add events for importing/exporting annotations/datasets (#7476)
This can provide useful statistics on the use of this functionality.

Note that I chose to use the same scope for operations on projects,
tasks and jobs, because the import/export operations work more-or-less
the same on each of them and have the same parameters. It's still
possible to distinguish which type of object the operation was performed
on by examining the various `*_id` fields in the event.