Fix several problems with the calculation of working time per client event
The first problem is that in the following case, the algorithm would accrue more
working time than was actually spent:

              A
        |-----------|
    ---------+------------+------> time
             B            C

A, B and C are events. Let's say that Te(A) is the timestamp of the end of A,
while T(B) and T(C) are the timestamps of B and C, respectively.

The current code in `ClientEventsSerializer.to_internal_value` adjusts
`last_timestamp` after processing every event. After A is processed,
`last_timestamp` is set to Te(A). After B is processed, `last_timestamp`
goes _backwards_ to T(B). So when the algorithm calculates the working time
for C, it gets T(C) - T(B), when the correct answer is T(C) - Te(A). The
span from T(B) to Te(A) gets counted twice.

Fix this by rewriting the algorithm, so that `last_timestamp` (now renamed
`previous_end_timestamp`) can only go forwards.

The second problem is that the algorithm is unable to calculate the working
time for the first event in each batch that the client sends. This is
because to calculate working time for an event, you need the
timestamp/duration of the previous event, and this information is unavailable
for the first event in the batch.

Fix this by resending the most recently sent event along with each batch,
and using it to initialize the algorithm.