Remove double `updated_date` bumping when saving annotations
Currently, `JobAnnotation.update` needlessly bumps the `updated_date` of the
job and the task twice (both via `_delete` and `_create`). These updates are
costly in terms of time due to the various side effects associated with them
(such as webhook sending and event logging), so it's better to avoid doing
them gratuitously.

Change the code so that the bumps only happen once.

I simplified the code a bit; previously, the code in `_delete` would only
bump the date if something was actually deleted (i.e., the annotations
passed in actually existed in the database). Now we just check if there were
any annotations we _wanted_ to delete. This means that if a request only
deletes non-existent annotations, the date will still get bumped. I don't
think it's a significant issue, since this is a pathological case.