Use the same algorithm to throttle quality reports as analytics reports
See 009f9f8a for more details. This patch extracts the high-level throttling
functionality added in that patch and reuses it for quality reports.

Note: in that patch I referred to this functionality as debouncing, but
throttling seems like a more accurate description. It would be debouncing
if the autoupdate job only ran after no updates occurred for a period,
which is not how it actually works.