Update python debugger configuration (#7634)
### Motivation and context
The old python debugger configuration will soon be deprecated so I
changed it to use debugpy which is the new configuration.

### Checklist
<!-- Go over all the following points, and put an `x` in all the boxes
that apply.
If an item isn't applicable for some reason, then ~~explicitly
strikethrough~~ the whole
line. If you don't do that, GitHub will show incorrect progress for the
pull request.
If you're unsure about any of these, don't hesitate to ask. We're here
to help! -->
- [x] I submit my changes into the `develop` branch
- [ ] I have created a changelog fragment <!-- see top comment in
CHANGELOG.md -->

### License

- [x] I submit _my code changes_ under the same [MIT License](
https://github.com/opencv/cvat/blob/develop/LICENSE) that covers the
project.
  Feel free to contact the maintainers if that's a concern.

Co-authored-by: syedbarimanjan <syedbarimnajan@gmail.com>
Co-authored-by: Maxim Zhiltsov <zhiltsov.max35@gmail.com>