Fix annotation accuracy display in the UI
The current formula the UI uses to calculate accuracy is incorrect. The denominator
is supposed to be the total sample count, but ds_count + gt_count - valid_count
is not the same value. To illustrate:

      ****      ***.        *...         ****
      ****      ***.        .*..         ****
      ****      ***.        ..*.         ****
      ....      ***.        ....         ***.

    ds_count  gt_count  valid_count  total_count (expected)

Each value is the sum of the values in the confusion matrix marked by `*`.
The current formula essentially adds values that are off the diagonal twice.

Instead of trying to derive the total count from other values, just transmit
the correct value from the server and use it in the UI.